requirements.txt
fastapi
uvicorn
python-dotenv
pymongo[srv]
fastapi[all]
selenium
undetected_chromedriver
bs4
requests